requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
pydantic==2.6.1
pydantic-settings==2.1.0
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
alembic==1.13.1
python-dotenv==1.0.0
email-validator==2.1.0
